Move parse_json_to_options to correct mod
