+++ native/src/internal_jni/io/write/csv.rs
@@ -6,7 +6,8 @@ use jni::JNIEnv;
 use jni_fn::jni_fn;
 use polars::prelude::*;
 
-use crate::internal_jni::io::write::{get_df_and_writer, parse_json_to_options};
+use crate::internal_jni::io::parse_json_to_options;
+use crate::internal_jni::io::write::get_df_and_writer;
 use crate::utils::error::ResultExt;
 
 #[jni_fn("org.polars.scala.polars.internal.jni.io.write$")]
